# Micro-Optimize Router#find_routes