--- a/actionpack/lib/action_dispatch/http/request.rb
+++ b/actionpack/lib/action_dispatch/http/request.rb
@@ -166,7 +166,7 @@ module ActionDispatch
     end
 
     def route=(route) # :nodoc:
-      set_header("action_dispatch.route", route)
+      @env["action_dispatch.route"] = route
     end
 
     def routes # :nodoc: