--- a/actionpack/test/journey/router/utils_test.rb
+++ b/actionpack/test/journey/router/utils_test.rb
@@ -18,12 +18,12 @@ module ActionDispatch
           assert_equal "a/b%20c+d%25?e", Utils.escape_fragment("a/b c+d%?e")
         end
 
-        def test_uri_unescape
-          assert_equal "a/b c+d", Utils.unescape_uri("a%2Fb%20c+d")
+        def test_CGI_unescapeURIComponent
+          assert_equal "a/b c+d", CGI.unescapeURIComponent("a%2Fb%20c+d")
         end
 
-        def test_uri_unescape_with_utf8_string
-          assert_equal "Šašinková", Utils.unescape_uri((+"%C5%A0a%C5%A1inkov%C3%A1").force_encoding(Encoding::US_ASCII))
+        def test_CGI_unescapeURIComponent_with_utf8_string
+          assert_equal "Šašinková", CGI.unescapeURIComponent((+"%C5%A0a%C5%A1inkov%C3%A1").force_encoding(Encoding::US_ASCII))
         end
 
         def test_normalize_path_not_greedy
@@ -36,7 +36,7 @@ module ActionDispatch
 
         def test_normalize_path_maintains_string_encoding
           path = "/foo%AAbar%AAbaz".b
-          assert_equal Encoding::ASCII_8BIT, Utils.normalize_path(path).encoding
+          assert_equal Encoding::BINARY, Utils.normalize_path(path).encoding
         end
 
         def test_normalize_path_with_nil