Followup: https://github.com/rails/rails/pull/54504
Followup: https://github.com/rails/rails/pull/54491
Followup: https://github.com/rails/rails/pull/54505
Followup: https://github.com/rails/rails/pull/54515
Followup: https://github.com/rails/rails/pull/54516

These are all small micro-optimizations, but they add up.

Same benchmark as: https://github.com/rails/rails/pull/54491#issuecomment-2650542281

```
== index ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    57.954k i/100ms
Calculating -------------------------------------
               after    589.889k (± 0.5%) i/s    (1.70 μs/i) -      2.956M in   5.010651s

Comparison:
              before:   448751.6 i/s
               after:   589888.6 i/s - 1.31x  faster

== show ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    46.760k i/100ms
Calculating -------------------------------------
               after    462.541k (± 1.0%) i/s    (2.16 μs/i) -      2.338M in   5.055185s

Comparison:
              before:   352192.0 i/s
               after:   462541.4 i/s - 1.31x  faster

== show_nested ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    42.082k i/100ms
Calculating -------------------------------------
               after    420.114k (± 0.2%) i/s    (2.38 μs/i) -      2.104M in   5.008430s

Comparison:
              before:   327672.8 i/s
               after:   420113.6 i/s - 1.28x  faster
```

If we compare to 8-0-stable to account for all the recent PRs, the overhead was cut by more than half:

```
== index ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    58.569k i/100ms
Calculating -------------------------------------
               after    590.724k (± 0.9%) i/s    (1.69 μs/i) -      2.987M in   5.056914s

Comparison:
              before:   272012.7 i/s
               after:   590724.1 i/s - 2.17x  faster

== show ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    46.784k i/100ms
Calculating -------------------------------------
               after    459.828k (± 0.5%) i/s    (2.17 μs/i) -      2.339M in   5.087265s

Comparison:
              before:   198611.2 i/s
               after:   459828.4 i/s - 2.32x  faster

== show_nested ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    41.386k i/100ms
Calculating -------------------------------------
               after    416.197k (± 0.5%) i/s    (2.40 μs/i) -      2.111M in   5.071520s

Comparison:
              before:   168787.4 i/s
               after:   416196.5 i/s - 2.47x  faster
```

Of course this is a micro-benchmark, at best you stand to save 4μs per request, which in the grand scheme of things is peanuts, but if that can appease the micro-benchmark crowd and help change perception, why not.

cc @skipkayhil